Adopted to aJson lib update

+++ lighthub/inputs.cpp
@@ -117,7 +117,7 @@ void Input::Parse(aJsonObject * configObj)
         itemBuffer = aJson.getObjectItem(inputObj, "S");
         if (!itemBuffer) {
             debugSerial<<F("In: ")<<pin<<F("/")<<inType<<endl;
-            aJson.addNumberToObject(inputObj, "S", 0);
+            aJson.addNumberToObject(inputObj, "S", (long int) 0);
             itemBuffer = aJson.getObjectItem(inputObj, "S");
         }
         if (itemBuffer) store = (inStore *) &itemBuffer->valueint;
